Titanic Project-Kaggle.py
I built a Titanic model that predicts the survival of individuals based on certain features. This model was developed using a neural network for a Kaggle competition. It achieves 88% accuracy on the test data.

Although the accuracy can be improved, the limited data provided by Kaggle restricts the model's potential for further learning. As you know, models become more mature and accurate with larger datasets. I also performed feature engineering to help the model learn more effectively.

If you have any queries, feel free to comment, and I'll respond as soon as possible.