Titanic Mdoel
Hey everyone!

I built a model using a neural network and performed feature engineering, which I found useful. The accuracy I achieved was nearly 88%, with a loss between 0.23 and 0.25. I tried reducing the loss and increasing the accuracy, but I found that the model was overfitting. Therefore, I concluded that this accuracy was the best I could achieve without overfitting.

The only thing you need to change when running the code is the path of the files being read the rest is good to go.

Link to Kaggle Notebook: https://www.kaggle.com/code/mushei/titanic-model